+++ mapEditor/src/panel/objects/bodyshape/BodyCompoundShapeWidget.cpp
@@ -211,14 +211,14 @@ namespace urchin {
             localizedShape->shape = bodyShapeWidget->retrieveShape();
 
             Vector3<float> eulerAngle(
-                    AngleConverter<float>::toRadian(eulerAxis0->value()),
-                    AngleConverter<float>::toRadian(eulerAxis1->value()),
-                    AngleConverter<float>::toRadian(eulerAxis2->value())
+                    AngleConverter<float>::toRadian((float)eulerAxis0->value()),
+                    AngleConverter<float>::toRadian((float)eulerAxis1->value()),
+                    AngleConverter<float>::toRadian((float)eulerAxis2->value())
             );
             QVariant variant = orientationType->currentData();
             auto rotationSequence = static_cast<Quaternion<float>::RotationSequence>(variant.toInt());
             localizedShape->transform = PhysicsTransform(
-                    Point3<float>(positionX->value(), positionY->value(), positionZ->value()),
+                    Point3<float>((float)positionX->value(), (float)positionY->value(), (float)positionZ->value()),
                     Quaternion<float>(eulerAngle, rotationSequence));
 
             localizedShapeTableView->updateSelectedLocalizedShape(localizedShape);
